moveRefactor

See merge request scycl11/platformer!66

+++ PlatformerPlain/src/main/java/com/example/platformerplain/move/Coord2D.java
@@ -1,10 +1,10 @@
 package com.example.platformerplain.move;
 
 public class Coord2D {
-    private float x;
-    private float y;
+    private double x;
+    private double y;
 
-    public Coord2D(float x, float y) {
+    public Coord2D(double x, double y) {
         this.x = x;
         this.y = y;
     }
@@ -14,37 +14,37 @@ public class Coord2D {
         this.y = 0;
     }
 
-    public float getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(float x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public float getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(float y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public float[] get() {
-        return new float[] {x, y};
+    public double[] get() {
+        return new double[] {x, y};
     }
 
-    public void set(float x, float y) {
+    public void set(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public void add(float dx, float dy) {
+    public void add(double dx, double dy) {
         this.x += dx;
         this.y += dy;
     }
 
-    public void reduce(float dx, float dy) {
+    public void reduce(double dx, double dy) {
         dx = Math.abs(dx);
         dy = Math.abs(dy);
         if (x > 0) {
@@ -60,24 +60,24 @@ public class Coord2D {
         }
     }
 
-    public void smoothReduce(float ax, float ay, float tx, float ty) {
-        float ratioX = 0, ratioY = 0;
-        if(tx != 0) ratioX = (float) Math.min(1.5, (Math.abs(x / tx)) + 0.25);
-        if(ty != 0) ratioY = (float) Math.min(1.5, (Math.abs(y / ty)) + 0.25);
+    public void smoothReduce(double ax, double ay, double tx, double ty) {
+        double ratioX = 0, ratioY = 0;
+        if(tx != 0) ratioX = Math.min(1.5, (Math.abs(x / tx)) + 0.25);
+        if(ty != 0) ratioY = Math.min(1.5, (Math.abs(y / ty)) + 0.25);
         reduce(ax * ratioX, ay * ratioY);
     }
 
-    public void smoothMinus(float ax, float ay, float tx, float ty) {
-        float ratioX = 0, ratioY = 0;
-        if(tx != 0) ratioX = (float) Math.min(1.5, (Math.abs(x / tx)) + 0.5);
-        if(ty != 0) ratioY = (float) Math.min(1.5, (Math.abs(y / ty)) + 0.5);
+    public void smoothMinus(double ax, double ay, double tx, double ty) {
+        double ratioX = 0, ratioY = 0;
+        if(tx != 0) ratioX = Math.min(1.5, (Math.abs(x / tx)) + 0.5);
+        if(ty != 0) ratioY = Math.min(1.5, (Math.abs(y / ty)) + 0.5);
         add(ax * ratioX, ay * ratioY);
     }
 
-    public void smoothAdd(float ax, float ay, float tx, float ty) {
-        float ratioX = 0, ratioY = 0;
-        if(tx != 0) ratioX = (float) Math.min(1.5, (Math.abs((x-tx) / tx)) + 0.25);
-        if(ty != 0) ratioY = (float) Math.min(1.5, (Math.abs((y-ty) / ty)) + 0.25);
+    public void smoothAdd(double ax, double ay, double tx, double ty) {
+        double ratioX = 0, ratioY = 0;
+        if(tx != 0) ratioX = Math.min(1.5, (Math.abs((x-tx) / tx)) + 0.25);
+        if(ty != 0) ratioY = Math.min(1.5, (Math.abs((y-ty) / ty)) + 0.25);
         add(ax * ratioX, ay * ratioY);
     }
 }
